don't go ready till we actually write out a conflist/syncnc
