+++ cns/healthserver/healthz.go
@@ -70,6 +70,7 @@ func NewHealthzHandlerWithChecks(cfg *Config) (http.Handler, error) {
 			return nil
 		}
 	}
+
 	return &healthz.Handler{
 		Checks: checks,
 	}, nil
